+++ src/render/icon/icon_overlay.h
@@ -4,7 +4,10 @@
 
 #pragma once
 
-#include <DirectXMath.h>
+#include "icon_geometry.h"
+
+#include <memory>
+#include <vector>
 #include <CommonStates.h>
 
 struct ID3D11Buffer;
@@ -17,12 +20,6 @@ struct ID3D11VertexShader;
 
 PLUGIN_NAMESPACE_BEGIN
 
-struct IconVertex
-{
-    DirectX::XMFLOAT3 pos;
-    DirectX::XMFLOAT4 color;
-};
-
 class IconOverlay
 {
 public:
@@ -31,14 +28,12 @@ public:
 
     bool init(ID3D11Device* device);
     void begin_frame(float width, float height);
-    void add_circle(float center_x, float center_y, float radius_px, DirectX::XMFLOAT4 const& color);
-    void add_triangle(float x0, float y0, float x1, float y1, float x2, float y2, DirectX::XMFLOAT4 const& a_color);
+    void add_marker(IconMarker const& marker, DirectX::XMFLOAT3 const& color);
     void draw(ID3D11DeviceContext* context, ID3D11RenderTargetView* target, std::shared_ptr<DirectX::DX11::CommonStates> const& states);
     void end_frame();
 private:
     bool create_pipeline(ID3D11Device* a_device);
     void release_pipeline();
-    void add_vertex_ndc(float a_ndc_x, float a_ndc_y, DirectX::XMFLOAT4 const& a_color);
 private:
     ID3D11VertexShader* m_vertex_shader = nullptr;
     ID3D11PixelShader* m_pixel_shader = nullptr;